# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_null_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
...
			"Problem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"statement": {
						"type": "string"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"inputFormat": {
						"type": "string"
					},
					"outputFormat": {
						"type": "string"
					},
					"solutionText": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					},
					"testcases": {
						"items": {
							"$ref": "#/components/schemas/Testcase"
						},
						"type": "array"
					},
					"isPublished": {
						"type": "boolean"
					},
					"associatedContests": {
						"items": {
							"$ref": "#/components/schemas/ProblemInContest"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"title",
					"statement",
					"difficulty",
					"timeLimit",
					"memoryLimit",
					"inputFormat",
					"outputFormat",
					"solutionText",
					"owner",
					"submissions",
					"testcases",
					"isPublished",
					"associatedContests"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Contest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"ruleText": {
						"type": "string"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"scoringRule": {
						"type": "string"
					},
					"isPlagiarismCheckEnabled": {
						"type": "boolean"
					},
					"isPublished": {
						"type": "boolean"
					},
					"organizer": {
						"$ref": "#/components/schemas/User"
					},
					"participations": {
						"items": {
							"$ref": "#/components/schemas/ContestParticipation"
						},
						"type": "array"
					},
					"problemsInContest": {
						"items": {
							"$ref": "#/components/schemas/ProblemInContest"
						},
						"type": "array"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"title",
					"description",
					"ruleText",
					"startTime",
					"endTime",
					"scoringRule",
					"isPlagiarismCheckEnabled",
					"isPublished",
					"organizer",
					"participations",
					"problemsInContest",
					"submissions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"GetUserResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_User.password_"
					}
				],
				"nullable": true
			},
			"ApiResponse_GetUserResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetUserResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
...
			"GetProblemResponse": {
				"$ref": "#/components/schemas/Problem"
			},
			"ApiResponse_GetProblemResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetProblemResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
...
			"GetAllContestsResponse": {
				"items": {
					"$ref": "#/components/schemas/Contest"
				},
				"type": "array"
			},
			"ApiResponse_GetAllContestsResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetAllContestsResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"GetContestResponse": {
				"$ref": "#/components/schemas/Contest"
			},
			"ApiResponse_GetContestResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetContestResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"GetContestRankingResponse": {
				"items": {
					"$ref": "#/components/schemas/ContestParticipation"
				},
				"type": "array"
			},
			"ApiResponse_GetContestRankingResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetContestRankingResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"GetContestProblemResponse": {
				"items": {
					"$ref": "#/components/schemas/Problem"
				},
				"type": "array"
			},
			"ApiResponse_GetContestProblemResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetContestProblemResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"DeleteProblemInContestRequest": {
				"properties": {
					"problemId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"problemId"
				],
				"type": "object"
			},
			"UpdateContestRequest": {
				"properties": {
					"isPublished": {
						"type": "boolean"
					},
					"isPlagiarismCheckEnabled": {
						"type": "boolean"
					},
					"scoringRule": {
						"type": "string"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"ruleText": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "web-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/user/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetUserResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/problem/id/{id}": {
			"get": {
				"operationId": "GetProblemById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetProblemResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "SoftDeleteProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/contest": {
			"get": {
				"operationId": "GetAllContests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetAllContestsResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateContest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"organizerId": {
										"type": "number",
										"format": "double"
									},
									"isPublished": {
										"type": "boolean"
									},
									"isPlagiarismCheckEnabled": {
										"type": "boolean"
									},
									"scoringRule": {
										"type": "string"
									},
									"endTime": {
										"type": "number",
										"format": "double"
									},
									"startTime": {
										"type": "number",
										"format": "double"
									},
									"ruleText": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"organizerId",
									"scoringRule",
									"endTime",
									"startTime",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/contest/{id}": {
			"get": {
				"operationId": "GetContestById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetContestResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteContest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "EditContest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateContestRequest"
							}
						}
					}
				}
			}
		},
		"/contest/search": {
			"get": {
				"operationId": "SearchContests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetAllContestsResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "searchKeywords",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startTimeLow",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startTimeHigh",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTimeLow",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endTimeHigh",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/contest/{id}/ranking": {
			"get": {
				"operationId": "GetContestRanking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetContestRankingResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/contest/{id}/problem": {
			"get": {
				"operationId": "GetContestProblems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetContestProblemResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteProblemInContestRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "EditProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"score": {
										"type": "number",
										"format": "double"
									},
									"problemId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"problemId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"post": {
				"operationId": "AddProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"score": {
										"type": "number",
										"format": "double"
									},
									"problemId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"problemId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/contest/{id}/editScore": {
			"patch": {
				"operationId": "EditScore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"score": {
										"type": "number",
										"format": "double"
									},
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"score",
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/contest/{id}/register": {
			"post": {
				"operationId": "RegisterToContest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}